+++ purchase_origin_address/README.rst
@@ -1,8 +1,7 @@
 Purchase Origin Address
 =======================
 
-This modules adds an origin address on the purchase order that will be
-transferred on picking in.
+This module was merged into purchase_addresses and will be dropped in v9.
 
 
 Contributors
@@ -10,3 +9,4 @@ Contributors
 
 * Yannick Vaucher <yannick.vaucher@camptocamp.com>
 * Leonardo Pistone <leonardo.pistone@camptocamp.com>
+* Alexandre Fayolle <alexandre.fayolle@camptocamp.com>
