+++ purchase_origin_address/__init__.py
@@ -1,2 +1,1 @@
 # -*- coding: utf-8 -*-
-from . import model
