+++ purchase_origin_address/__openerp__.py
@@ -29,9 +29,7 @@
  'depends': ['purchase',
              ],
  'demo': [],
- 'data': ['view/purchase_order.xml',
-          'view/stock_picking.xml',
-          ],
+ 'data': [],
  'auto_install': False,
  'test': [],
  'installable': True,
