--- purchase_origin_address/model/__init__.py
@@ -1,3 +1,0 @@
-# -*- coding: utf-8 -*-
-from . import purchase_order
-from . import stock_picking
