--- purchase_origin_address/view/purchase_order.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0"?>
-<openerp>
-  <data>
-
-    <record model="ir.ui.view" id="purchase_order_form">
-      <field name="name">purchase.order.inherit</field>
-      <field name="model">purchase.order</field>
-      <field name="inherit_id" ref="purchase.purchase_order_form"/>
-      <field name="arch" type="xml">
-
-        <field name="partner_id" position="after">
-          <field name="origin_address_id"/>
-        </field>
-
-      </field>
-    </record>
-
-  </data>
-</openerp>
