--- purchase_origin_address/view/stock_picking.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0"?>
-<openerp>
-  <data>
-
-    <record id="view_picking_form" model="ir.ui.view">
-      <field name="name">stock.picking.form</field>
-      <field name="model">stock.picking</field>
-      <field name="inherit_id" ref="stock.view_picking_form"/>
-      <field name="arch" type="xml">
-
-        <field name="partner_id" position="after">
-          <field name="origin_address_id"/>
-        </field>
-
-      </field>
-    </record>
-
-  </data>
-</openerp>
